replaced Main with MainWithPublish
